[ghcui]集成货物详情接口

+++ FengLe/app/src/main/java/com/yunqi/fengle/model/bean/CustomerContactDetail.java
@@ -1,16 +1,14 @@
 package com.yunqi.fengle.model.bean;
 
-import java.io.Serializable;
+import com.google.gson.annotations.SerializedName;
 
 /**
  * @author ghcui
  * @time 2017/2/15
  */
 public class CustomerContactDetail extends Customer{
-    public int fahuo_num;//发货数
-    public int huikuan_num;//回款
-    public int xinren_bili;//信任比率
-    public int real_bili;//实际比率
-
-
+    public float fahuo_num;//发货数
+    public float huikuan_num;//回款
+    public float xinren_bili;//信任比率
+    public float real_bili;//实际比率
 }
